move acomsisCookieHandler into utils.js
